Roles and privilages added. classes created for attendance comparison report

+++ src/main/java/com/outreach/greenstar/service/RolesService.java
@@ -1,7 +1,6 @@
 package com.outreach.greenstar.service;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 
@@ -9,16 +8,20 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.outreach.greenstar.dao.RolesDao;
+import com.outreach.greenstar.dto.PrivilagesDTO;
 import com.outreach.greenstar.dto.RoleDTO;
+import com.outreach.greenstar.entities.Privilages;
 import com.outreach.greenstar.entities.Role;
+import com.outreach.greenstar.repository.PrivilagesRepository;
 import com.outreach.greenstar.utility.EntityDtoConverter;
-import com.outreach.greenstar.utility.PrivilageEnum;
 
 @Service("rolesService")
 public class RolesService {
 
     @Autowired
     private RolesDao rolesDao;
+    @Autowired
+    private PrivilagesRepository privilagesRepository;
     
     public List<RoleDTO> getAllRoles() {
         List<Role> listOfRoles = rolesDao.getAllRoles();
@@ -37,6 +40,14 @@ public class RolesService {
 
     public RoleDTO createRole(RoleDTO rolesDTO) {
         Role role = EntityDtoConverter.getRole(rolesDTO);
+        List<PrivilagesDTO> privilages = rolesDTO.getPrivilages();
+        List<Privilages> newPrivilages = new ArrayList<>();
+        for (Iterator<PrivilagesDTO> iterator = privilages.iterator(); iterator.hasNext();) {
+            PrivilagesDTO privilages2 = iterator.next();
+            Privilages priv = privilagesRepository.getOne(privilages2.getId());
+            newPrivilages.add(priv);
+        }
+        role.setListOfPrivilages(newPrivilages);
         role = rolesDao.saveOrUpdate(role);
         return EntityDtoConverter.getRolesDto(role);
     }
@@ -46,12 +57,17 @@ public class RolesService {
         return createRole(rolesDTO);
     }
 
-    public List<String> getAllPrivilages() {
-        return Arrays.asList(PrivilageEnum.stringValues());
+    public List<Privilages> getAllPrivilages() {
+        return rolesDao.getAllPrivilages();
     }
 
     public void deleteRole(int roleId) {
         Role role = rolesDao.getRoleById(roleId);
+        if (role.getRoleName().equalsIgnoreCase("admin")) {
+            throw new IllegalArgumentException("Cannot delete Admin user");
+        }
+        role.getListOfPrivilages().clear();
+        rolesDao.saveOrUpdate(role);
         rolesDao.deleteRole(role);
     }
